Address code review feedback: clean up implementation for bremsstrahlung and Mavrin validation
